+++ lib/graphql_rails/controller/action_configuration.rb
@@ -31,14 +31,13 @@ module GraphqlRails
         self
       end
 
-      def permit_input(name, type: nil, description: nil, subtype: nil)
+      def permit_input(name, type: nil, options: {}, **input_options)
         field_name = name.to_s.remove(/!\Z/)
 
         attributes[field_name] = Attributes::InputAttribute.new(
           name.to_s, type,
-          description: description,
-          subtype: subtype,
-          options: action_options
+          options: action_options.merge(options),
+          **input_options
         )
         self
       end